split the logfilter source
