+++ consensus/dccs/2_dccs.go
@@ -19,18 +19,11 @@ package dccs
 
 import (
 	"bytes"
-	"context"
 	"errors"
 	"fmt"
 	"math/big"
 	"sort"
 	"time"
-
-	"github.com/ethereum/go-ethereum/core/rawdb"
-
-	"github.com/ethereum/go-ethereum/eth/filters"
-	"github.com/ethereum/go-ethereum/ethdb"
-	"github.com/ethereum/go-ethereum/rpc"
 
 	"github.com/ethereum/go-ethereum/accounts"
 	"github.com/ethereum/go-ethereum/common"
@@ -231,46 +224,6 @@ func (d *Dccs) verifySeal2(chain consensus.ChainReader, header *types.Header, pa
 	return nil
 }
 
-type LogFilterBackend struct {
-	chain consensus.ChainReader
-	db    ethdb.Reader
-}
-
-func (b *LogFilterBackend) HeaderByNumber(ctx context.Context, blockNr rpc.BlockNumber) (*types.Header, error) {
-	return b.chain.GetHeaderByNumber(uint64(blockNr.Int64())), nil
-}
-
-func (b *LogFilterBackend) HeaderByHash(ctx context.Context, blockHash common.Hash) (*types.Header, error) {
-	return b.chain.GetHeaderByHash(blockHash), nil
-}
-
-func (b *LogFilterBackend) GetReceipts(ctx context.Context, blockHash common.Hash) (types.Receipts, error) {
-	header := b.chain.GetHeaderByHash(blockHash)
-	if header == nil {
-		return nil, nil
-	}
-	receipts := rawdb.ReadReceipts(b.db, blockHash, header.Number.Uint64(), b.chain.Config())
-	return receipts, nil
-}
-
-func (b *LogFilterBackend) GetLogs(ctx context.Context, blockHash common.Hash) ([][]*types.Log, error) {
-	receipts, _ := b.GetReceipts(ctx, blockHash)
-	logs := make([][]*types.Log, len(receipts))
-	for i, receipt := range receipts {
-		logs[i] = receipt.Logs
-	}
-	return logs, nil
-}
-
-// This nil assignment ensures compile time that LogFilterBackend implements filters.SimpleBackend.
-var _ filters.SimpleBackend = (*LogFilterBackend)(nil)
-
-// Keccak256("Joined(address,address)")
-var joinedTopic = common.HexToHash("7702dccda75540ad1dca8d5276c048f4a5c0e4203f6da4be214bfb1901b203ea")
-
-// Keccak256("Left(address,address)")
-var leftTopic = common.HexToHash("4b9ee4dd061ba088b22898a02491f3896a4a580c6cda8783ca579ee159f8e8c5")
-
 func isAuthorized(signer common.Address, sealers []common.Address) bool {
 	for _, sealer := range sealers {
 		if signer == sealer {
@@ -391,46 +344,6 @@ func (d *Dccs) getSealers(number uint64, chain consensus.ChainReader, parents []
 	}
 	log.Error("getSealers", "sealers", list)
 	return list, nil
-}
-
-// fetchSealerApplications filters the block for any joining or leaving sealer.
-// Multiple sealer applications can be confirmed in the same block, the order of
-// the requests kept as is.
-func (d *Dccs) fetchSealerApplications(header *types.Header, chain consensus.ChainReader) ([]sealerApplication, error) {
-	logs, err := filters.BlockLogs(header,
-		[]common.Address{d.config.Contract},
-		[][]common.Hash{{joinedTopic, leftTopic}},
-		&LogFilterBackend{
-			chain: chain,
-			db:    d.db,
-		})
-
-	if err != nil {
-		return nil, err
-	}
-	if len(logs) == 0 {
-		return nil, nil
-	}
-
-	applications := make([]sealerApplication, len(logs))
-
-	for i, l := range logs {
-		// len(log.Data) must be 32 * 2 here
-		sealer := common.BytesToAddress(l.Data[32:])
-		var joined bool
-		if l.Topics[0] == joinedTopic {
-			joined = true
-			staker := common.BytesToAddress(l.Data[:32])
-			log.Error("Sealer joined", "sealer", sealer, "coinbase", staker)
-		} else {
-			log.Error("Sealer left", "sealer", sealer)
-		}
-		applications[i] = sealerApplication{
-			sealer: sealer,
-			action: joined,
-		}
-	}
-	return applications, nil
 }
 
 func (d *Dccs) isSealerApplicationBlock(header *types.Header) bool {
